feat:implement panel add button
