--- src/pages/C5th.js
@@ -1,75 +1,0 @@
-// material
-import { Box, Grid, Container, Typography } from '@material-ui/core';
-// components
-import Page from '../components/Page';
-import {
-  AppTasks,
-  AppNewUsers,
-  AppBugReports,
-  AppItemOrders,
-  AppNewsUpdate,
-  AppWeeklySales,
-  AppOrderTimeline,
-  AppTrafficBySite,
-  TestPanel
-} from '../components/_dashboard/app';
-
-// ----------------------------------------------------------------------
-
-export default function C5th() {
-  return (
-    <Page title="Dashboard | C5th">
-      <Container maxWidth="xl">
-        <Box sx={{ pb: 5 }}>
-          <Typography variant="h4">Hi, Welcome back</Typography>
-        </Box>
-        <Grid container spacing={3}>
-          <Grid item xs={12} sm={6} md={3}>
-            <AppWeeklySales />
-          </Grid>
-          <Grid item xs={12} sm={6} md={3}>
-            <AppNewUsers />
-          </Grid>
-          <Grid item xs={12} sm={6} md={3}>
-            <AppItemOrders />
-          </Grid>
-          <Grid item xs={12} sm={6} md={3}>
-            <AppBugReports />
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={8}>
-            <TestPanel />
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={4}>
-            <div>test</div>
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={8}>
-            <div>test</div>
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={4}>
-            <div>test</div>
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={8}>
-            <AppNewsUpdate />
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={4}>
-            <AppOrderTimeline />
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={4}>
-            <AppTrafficBySite />
-          </Grid>
-
-          <Grid item xs={12} md={6} lg={8}>
-            <AppTasks />
-          </Grid>
-        </Grid>
-      </Container>
-    </Page>
-  );
-}
